% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick box plot in ggplot.}
\usage{
gbox(x, y = NULL, horizontal = FALSE, box_color = "black")
}
\arguments{
\item{x}{This is the vector to be box plotted.}

\item{y}{This is a categorical vector optionally supplied for side by side plots.
The default is NULL.}

\item{horizontal}{This is a True or False variable that determines whether the
plots are horizontal or vertical. The default is FALSE.}

\item{box_color}{This is the color of the box plot that will be plotted.
The default is color black.}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This will graph a vector in a ggplot-style box plot with the
x-axis labeled "x".
}
\examples{
## Create a box plot of x.
x <- rnorm(100)
gbox(x, horizontal = TRUE, box_color = 'blue')

## Create a box plot of x
x <- rnorm(100)
y <- factor(rep(c("A", "B"), each = 50))
gbox(x,y)

}
